pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-bridge-pipeline"
version = "0.1.0"
description = "Manifest-driven data bridge pipeline with reconcile and safe queue semantics"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Open Source Contributors"}]
dependencies = [
  "typer>=0.12",
  "pydantic>=2.6",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "black>=24.0",
  "mypy>=1.10",
]

[project.scripts]
data-bridge-pipeline = "data_bridge_pipeline.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
